requirements.txt
flask==3.0.2
google-generativeai
python-dotenv==1.0.1
nltk==3.8.1 